Place orders. Start work on deliveries

+++ db/migrate/20140920225110_add_rating_to_order.rb
@@ -1,5 +1,5 @@
 class AddRatingToOrder < ActiveRecord::Migration
   def change
-    add_column :orders, :rating, :integer
+    add_column :orders, :rating, :decimal
   end
 end
